Converted weight to float32.

+++ pagerank.go
@@ -8,27 +8,27 @@ import (
 )
 
 type node struct {
-	weight   float64
-	outbound float64
+	weight   float32
+	outbound float32
 }
 
 // Graph holds node and edge data.
 type Graph struct {
-	edges map[uint32](map[uint32]float64)
+	edges map[uint32](map[uint32]float32)
 	nodes map[uint32]*node
 }
 
 // NewGraph initializes and returns a new graph.
 func NewGraph() *Graph {
 	return &Graph{
-		edges: make(map[uint32](map[uint32]float64)),
+		edges: make(map[uint32](map[uint32]float32)),
 		nodes: make(map[uint32]*node),
 	}
 }
 
 // Link creates a weighted edge between a source-target node pair.
 // If the edge already exists, the weight is incremented.
-func (self *Graph) Link(source, target uint32, weight float64) {
+func (self *Graph) Link(source, target uint32, weight float32) {
 	if _, ok := self.nodes[source]; ok == false {
 		self.nodes[source] = &node{
 			weight:   0,
@@ -46,7 +46,7 @@ func (self *Graph) Link(source, target uint32, weight float64) {
 	}
 
 	if _, ok := self.edges[source]; ok == false {
-		self.edges[source] = map[uint32]float64{}
+		self.edges[source] = map[uint32]float32{}
 	}
 
 	self.edges[source][target] += weight
@@ -57,9 +57,9 @@ func (self *Graph) Link(source, target uint32, weight float64) {
 // ε (epsilon) is the convergence criteria, usually set to a tiny value.
 //
 // This method will run as many iterations as needed, until the graph converges.
-func (self *Graph) Rank(α, ε float64, callback func(id uint32, rank float64)) {
-	Δ := float64(1.0)
-	inverse := 1 / float64(len(self.nodes))
+func (self *Graph) Rank(α, ε float32, callback func(id uint32, rank float32)) {
+	Δ := float32(1.0)
+	inverse := 1 / float32(len(self.nodes))
 
 	// Normalize all the edge weights so that their sum amounts to 1.
 	for source := range self.edges {
@@ -75,8 +75,8 @@ func (self *Graph) Rank(α, ε float64, callback func(id uint32, rank float64)) 
 	}
 
 	for Δ > ε {
-		leak := float64(0)
-		nodes := map[uint32]float64{}
+		leak := float32(0)
+		nodes := map[uint32]float32{}
 
 		for key, value := range self.nodes {
 			nodes[key] = value.weight
@@ -101,7 +101,7 @@ func (self *Graph) Rank(α, ε float64, callback func(id uint32, rank float64)) 
 		Δ = 0
 
 		for key, value := range self.nodes {
-			Δ += math.Abs(value.weight - nodes[key])
+			Δ += float32(math.Abs(float64(value.weight - nodes[key])))
 		}
 	}
 
@@ -112,6 +112,6 @@ func (self *Graph) Rank(α, ε float64, callback func(id uint32, rank float64)) 
 
 // Reset clears all the current graph data.
 func (self *Graph) Reset() {
-	self.edges = make(map[uint32](map[uint32]float64))
+	self.edges = make(map[uint32](map[uint32]float32))
 	self.nodes = make(map[uint32]*node)
 }
